pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "q4-rescue"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "psycopg[binary]>=3.2",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "python-dotenv>=1.0",
  "prometheus-client",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-exporter-otlp-proto-http"
]

[project.optional-dependencies]
test = [
  "httpx>=0.27"
]

[tool.setuptools.packages.find]
include = ["app*"]
